+++ CHANGELOG.md
@@ -5,6 +5,13 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- `capture` and `restore` are now hoisted functions.
+- Fixed timing issues with radio buttons and validation with side-effects.
+
 ## [0.8.4] - 2023-04-27
 
 ### Fixed